requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-autoslug==1.9.9
django-crispy-forms==2.0
django-tinymce==3.6.1
Pillow==9.5.0
slugify==0.0.1
sqlparse==0.4.3
tzdata==2023.3
